requirements.txt
# Core dependencies
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.18.0
tenacity>=8.2.0
rich>=13.0.0
jsonschema>=4.17.0

# Performance tracking and visualization
matplotlib>=3.7.0

# Manifold bot framework
manifoldbot>=0.1.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
